requirements.txt
torch
torchvision
pillow
tqdm
matplotlib
pytorch-ignite
scipy
